os/darwin: remove code that changed the device class used
To support macOS El Capitan I added some code to change from using the
IOUSBDevice class to using IOUSBHostDevice. This looked necessary as
Apple had re-writed IOUSBFamily and IOUSBHostDevice was the new base
class name. The problem is that using the new class misses some
devices in more recent versions of macOS. Apple is still including the
IOClassNameOverride property (as of macOS Mojave) so it is still safe
to search using the IOUSBDevice class. We will need to find a
different class if Apple ever removes the IOClassNameOverride
property from USB devices.

Closes #428

Signed-off-by: Nathan Hjelm <hjelmn@mac.com>