linux: support non-hotplug device discovery
The non-hotplug path is still useful to some non-
privileged containers and VMs that cannot receive
uevents via the netlink because they are restricted
for good reasons like security.

Without uevents, the existing applications cannot
get connection changes after an initial reading in
these containers.

This issue is not exactly caused by libusb. But
from a system view across kernel and user space,
support this by reusing the cold non-hotplug path
seems a good remedy without any penalty.

We could have a further change to make the config
as a runtime option so a recompiling in container can
be saved.

Note, this change doesn't make apps gain access to
devices. What it resolves is the discovery part.

Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>
Signed-off-by: Your Name <you@example.com>