fix issues in libusb_exit()
There are several issues in the same section of API
libusb_exit() so we have a single patch to fix all
of them.

() With the hotplug, the existing leaking check is
not effective. The device list of the context is
always removed before the check, so even an app
forgot to unref a device or free the device list,
this check doesn't catch such bugs.

() For the non-hotplug (still useful to the
further usages), cleanup is not performed.

() The existing libusb_unref_device() calls
usbi_disconnect_device(). The later tries to grab
ctx->usb_devs_lock, so we cannot use them when
iterating the device list with holding the same lock
at an outer level of calling (deadlock). A new added
helper function is added for core and backends to
handle these special situations.

Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>
Signed-off-by: Your Name <you@example.com>