Add not implemented guard for linux like systems
When compiling for android and disable udev the following functions do
nothing: linux_scan_devices() and op_hotplug_poll(). The result is that
libusb is not able to list devices.

This patch raise a compile time error if the config.h results in no
implementation for those functions.

For instance if USE_UDEV is not defined and __ANDROID__ is defined the
following error will be raised "<function> not implemented for this config".

Signed-off-by: Vinicius Tinti <vinicius.tinti@almg.gov.br>