linux_udev: silently ignore "bind" action
When a driver is bound to a device, udev emits a "bind" action that
causes libusb to log an error message:

	libusb: error [udev_hotplug_event] ignoring udev action bind

Since we know this action is not relevant for libusb, silently ignore it
to avoid people thinking there is something wrong.

There is already a debug log entry for the action, so there is no need
to add a duplicate here.

Signed-off-by: John Keeping <john@metanate.com>