Windows: Fix poll implementation for NDEBUG (Release) builds
The refactoring in commit df61c0c3a3 ("Windows: improve poll
abstraction") introduced a bug in builds where NDEBUG is defined because
of a statement with side-effects that was put inside an assertion. When
this statement is not evaluated, the file descriptor table gets corrupt.
Fix this by moving the statement outside of the assertion.

Signed-off-by: Chris Dickens <christopher.a.dickens@gmail.com>