core: Use a consistent variable name for usbi_transfer structures
Most places in the library use the name 'itransfer' when referring to a
usbi_transfer structure. This is helpful to distinguish between the
public libusb_transfer structure and the internal structure. Fix up the
few places that don't follow this convention so that it is consistent
across the entire library.

Signed-off-by: Chris Dickens <christopher.a.dickens@gmail.com>