core: Fix return value of usbi_clock_gettime()
In most cases, usbi_clock_gettime() will map to the standard library's
clock_gettime() function. The semantics of this function are that it
returns -1 upon failure with the error code available in the errno
variable. The backends that need to implement this function should
follow the same semantics, and the return value of usbi_clock_gettime()
should not be directly propagated upwards.

Signed-off-by: Chris Dickens <christopher.a.dickens@gmail.com>